requirements.txt
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
six==1.16.0
SpeechRecognition==3.10.0
spotipy==2.23.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
